POC Make UI compatible with the new srvid column in powa4
A new screen is added as the first screen to choose the server to display.

All datasources now use an additional qual "srvid = ?" to get the data related
to the chosen server.

Also teach the queries that needs to connect on the real database (qual
resolution, explain, wizard...) to do so.

powa-remote daemon is also updated to connect to the repository and get the
list of servers for powa_servers instead of its own config file.